# bookstore-schemas/schemas/openapi.yaml
openapi: 3.0.4
info:
  title: Book store
  description: Книжный магазин
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
paths:
...
  /account/purchased-books:
    get:
      tags:
        - account

      summary: Список приобретенных книг
      description: Получение списка приобретенных книг
      operationId: getPurchasedBooks

      security:
        - bearerAuth: []

      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/SortBook'
        - $ref: '#/components/parameters/Order'

        - in: query
          name: accountId
          description: ID аккаунта
          schema:
            type: integer

      responses:
        '200':
          description: Список купленных книг

          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Book'

        '4XX':
          description: Клиентская ошибка
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '5XX':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /account/favorite:
    post:
      tags:
        - account

      summary: Добавить книгу в избранное
      description: Добавление книги в избранное
      operationId: addFavoriteBook

      parameters:
        - in: query
          name: accountId
          description: ID аккаунта
          schema:
            $ref: '#/components/schemas/account-id'

        - $ref: '#/components/parameters/BookId'


      responses:
        '201':
          description:  Книга добавлена в избранное

          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Book'

        '4XX':
          description: Клиентская ошибка
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '5XX':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      tags:
        - account

      summary: Список книг в избранном
      description: Получение информации о списке избранных книг
      operationId: getFavoriteBooks

      parameters:
        - in: query
          name: accountId
          description: ID аккаунта
          schema:
            $ref: '#/components/schemas/account-id'

        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/SortFavorite'
        - $ref: '#/components/parameters/Order'

      responses:
        '200':
          description: Список книг в избранном

          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Book'

        '4XX':
          description: Клиентская ошибка
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '5XX':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /account/basket:
    get:
      tags:
        - account

      summary: Корзина с книгами
      description: Получение корзины с деталями
      operationId: getBasket

      security:
        - bearerAuth: []

      parameters:
        - in: query
          name: accountId
          description: ID аккаунта
          schema:
            type: integer

      responses:
        '200':
          description: Корзина с деталями

          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Basket'

        '4XX':
          description: Клиентская ошибка
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '5XX':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /account/basket/items:
    post:
      tags:
        - account

      summary: Добавить книгу в корзину
      description: Добавление книги в корзину
      operationId: addToBasket

      security:
        - bearerAuth: []

      parameters:
        - in: query
          name: accountId
          description: ID аккаунта
          schema:
            type: integer


      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookIdRequest'

      responses:
        '201':
          description: Книга добавлена в корзину
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Basket'
        '4XX':
          description: Клиентская ошибка
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '5XX':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /account/basket/items/{itemId}:
    delete:
      tags:
        - account
      summary: Удаление товара из корзины
      description: Удаление товара из корзины по его id
      operationId: removeBasketItem

      security:
        - bearerAuth: []

      parameters:
        - name: itemId
          in: path
          required: true
          description: ID позиции корзины
          schema:
            $ref: '#/components/schemas/item-id'

      responses:
        '204':
          description: Удалено из корзины
        '4XX':
          description: Клиентская ошибка
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '5XX':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /auth/register:
    post:
      tags:
        - auth

      summary: Регистрация пользователя
      description: Осуществление регистрации пользователя
      operationId: registerUser

      requestBody:
        required: true

        content:
          application/json:
            schema:
              type: object

              required:
                - email
                - password

              properties:
                email:
                  type: string

                password:
                  type: string

      responses:
        '201':
          description: Пользователь зарегистрирован
        '4XX':
          description: Клиентская ошибка
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '5XX':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/login:
    post:
      tags:
        - auth

      summary: Ввод данных пользователя
      description: Авторизация пользователя по email и паролю
      operationId: loginUser

      requestBody:
        required: true

        content:
          application/json:
            schema:
              type: object

              required:
                - email
                - password

              properties:
                email:
                  type: string
                  format: email
                  description: Email пользователя

                password:
                  type: string
                  description: Пароль пользоватея

      responses:
        '200':
          description: Успешная авторизация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginUserResponse'

        '4XX':
          description: Клиентская ошибка
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '5XX':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/logout:
    post:
      tags:
        - auth

      summary: Выход из профиля
      description: Осуществление выхода из профиля
      operationId: postLogout

      security:
        - bearerAuth: []

      responses:
        '200':
          description: Сессия удалена

        '4XX':
          description: Клиентская ошибка
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '5XX':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /review:
    post:
      tags:
        - review

      summary: Отзыв на книгу
      description: Отзыв на книгу по id книги
      operationId: postReview

      security:
        - bearerAuth: []

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookReviewInput'

      responses:
        '201':
          description:  Отзыв на книгу добавлен

          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookReview'

        '4XX':
          description: Клиентская ошибка
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '5XX':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /review/{reviewId}:
    get:
      tags:
        - review

      summary: Отзыв на книгу
      description: Метод для получения отзыва на книгу по id
      operationId: getBookReviewById

      parameters:
        - name: reviewId
          in: path
          required: true
          description: Параметр для идентификации по id для поиска отзыва
          schema:
            $ref: '#/components/schemas/review-id'

      responses:
        '200':
          description: Вывод отзыва по id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookReview'

        '4XX':
          description: Клиентская ошибка
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '5XX':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      tags:
        - review

      summary: Внесение изменений в отзыв
      description: Внесение изменений по id отзыва
      operationId: changeReview

      security:
        - bearerAuth: []

      parameters:
        - name: reviewId
          in: path
          required: true
          description: Параметр для идентификации по id для изменения отзыва
          schema:
            $ref: '#/components/schemas/review-id'

      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                text:
                  type: string
                rating:
                  type: integer

      responses:
        '200':
          description: Отзыв на книгу изменен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookReview'
        '4XX':
          description: Клиентская ошибка
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '5XX':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - review

      summary: Удаление отзыва
      description: Удаление отзыва по его ID
      operationId: deleteReviewById

      security:
        - bearerAuth: []

      parameters:
        - name: reviewId
          in: path
          required: true
          description: ID отзыва
          schema:
            $ref: '#/components/schemas/review-id'

      responses:
        '204':
          description: Отзыв успешно удалён

        '4XX':
          description: Клиентская ошибка
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '5XX':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /reviews:
    get:
      tags:
        - review

      summary: Список отзывов
      description: Вывод списка отзывов
      operationId: getReviews

      parameters:
        - in: query
          name: bookId
          description: ID книги
          schema:
            type: integer

        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/SortReview'
        - $ref: '#/components/parameters/Order'

      responses:
        '200':
          description: Список отзывов по книге

          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BookReview'

        '4XX':
          description: Клиентская ошибка
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '5XX':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    BookId:
      name: bookId
      in: query
      description: Идентификатор книги
      required: true
      schema:
        type: integer
    Limit:
      name: limit
      in: query
      description: Количество элементов
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
    Offset:
      name: offset
      in: query
      description: Смещение
      schema:
        type: integer
        minimum: 0
        default: 0
    SortBook:
      name: sortBook
      in: query
      description: Поле сортировки
      schema:
        type: string
        enum:
          - id
          - title
          - price
    SortReview:
      name: sortBook
      in: query
      description: Поле сортировки
      schema:
        type: string
        enum:
          - id
          - reviewerId
          - rating
    SortFavorite:
      name: sortBook
      in: query
      description: Поле сортировки
      schema:
        type: string
        enum:
          - id
          - createdAt
    Order:
      name: order
      in: query
      description: Направление сортировки
      schema:
        type: string
        enum:
          - asc
          - desc
        default: asc
  schemas:
    book-id:
      type: integer
      description: Идентификатор книги
    account-id:
      type: integer
      description: Идентификатор аккаунта
...
    item-id:
      type: integer
      description: Идентификатор корзины
    author-id:
      type: integer
      description: Идентификатор автора
    review-id:
      type: integer
      description: Идентификатор отзыва
    publisher-id:
      type: integer
      description: Идентификатор публикатора
    publisherSeries-id:
      type: integer
      description: Идентификатор серии
    Book:
      type: object

      properties:
        id:
          $ref: '#/components/schemas/book-id'

        title:
          type: string
          description: Название книги

        author:
          $ref: '#/components/schemas/Author'

        genre:
          $ref: '#/components/schemas/Genre'

        price:
          type: number
          description: Информация о цене

        publisherId:
          $ref: '#/components/schemas/publisher-id'

        ISBN:
          type: string
          description: Информация о муждународном книжном номере

        series:
          $ref: '#/components/schemas/PublisherSeries'

        release:
          type: string
          description: Информация о дате выпуска книги

        release-date:
          type: string
          format: date-time

        description:
          type: string
          description: Описание книги

        pages:
          type: integer
          description: Информация о количестве страниц

        weight:
          type: number
          format: double
          description: Информация о весе книги

        reviewCount:
          type: integer
          description: Количество отзывов

        averageRating:
          type: number
          description: Средний рейтинг
    Author:
      type: object

      properties:
        id:
          $ref: '#/components/schemas/author-id'

        name:
          type: string
          description: Имя автора

        surname:
          type: string
          description: Описание книги

        middleName:
          type: string
          description: Отчество автора

        pseudonym:
          type: string
          description: Псевдоним автора
    Genre:
      type: string
      enum:
        - HORROR
        - COMEDY
        - ROMANTIC
        - DRAMA
        - THRILLER
    Basket:
      type: object

      properties:
        id:
          type: integer
          description: Идентификатор корзины

        basketDetails:
          type: array
          description: Детализация информации о книгах в корзине
          items:
            $ref: '#/components/schemas/BasketDetail'

        totalPrice:
          type: number
          description: Итоговая стоимость книг

        quantityBooks:
          type: integer
          description: Количество книг

        createdAt:
          type: string
          format: date-time
          description: Информация о дате создания корзины

        updatedAt:
          type: string
          format: date-time
    BasketDetail:
      type: object
      description: Позиция книги в корзине

      properties:
        id:
          $ref: '#/components/schemas/item-id'

        bookId:
          $ref: '#/components/schemas/book-id'

        title:
          type: string
          description: Название книги

        price:
          type: number
          description: Цена одной книги

        quantity:
          type: integer
          description: Количество экземпляров книги в корзине

        totalPrice:
          type: number
          description: Общая стоимость позиции (price * quantity)
...
    PublisherSeries:
      type: object

      properties:
        id:
          $ref: '#/components/schemas/publisherSeries-id'

        name:
          type: string
          description: Информация о названии серии

        publisherId:
          $ref: '#/components/schemas/publisher-id'

        createdAt:
          type: string
          format: date-time
          description: Информация о дате создания серии

        updatedAt:
          type: string
          format: date-time
          description: Информация о дате одновления серии
...
    BookReview:
      type: object

      properties:
        id:
          $ref: '#/components/schemas/review-id'

        book:
          $ref: '#/components/schemas/Book'

        text:
          type: string
          description: Текст отзыва

        reviewerId:
          type: integer
          description: Идентификатор пользователя, оставившего отзыв

        rating:
          type: integer
          description: Информация о рейтинге фильма
          minimum: 1
          maximum: 5
          default: 5

        createdAt:
          type: string
          format: date-time
          description: Информация о дате сосздания отзыва

        updatedAt:
          type: string
          format: date-time
          description: Информация о дате обновления отзыва
    BookReviewInput:
      type: object
      required:
        - bookId
        - text
        - rating
        - reviewerId
      properties:
        bookId:
          $ref: '#/components/schemas/book-id'
        text:
          type: string
          description: Текст отзыва
        rating:
          $ref: '#/components/schemas/BookReview/properties/rating'
        reviewerId:
          type: integer
          description: Идентификатор пользователя, оставившего отзыв
...
    UserDto:
      type: object
      properties:
        id:
          type: integer
          description: Идентификатор пользователя
        email:
          type: string
          description: Email пользователя
        name:
          type: string
          description: Имя
        surname:
          type: string
          description: Фамилия
        middleName:
          type: string
          description: Отчество
        phoneNumber:
          type: string
          description: Номер телефона
    LoginUserResponse:
      type: object
      properties:
        accessToken:
          type: string
          description: JWT токен для аутентификации
        refreshToken:
          type: string
          description: Refresh токен для обновления accessToken
        user:
          $ref: '#/components/schemas/UserDto'
    BookIdRequest:
      type: object
      required:
        - bookId
      properties:
        bookId:
          type: integer
    Error:
      type: object

      properties:
        message:
          type: string